requirements.txt
# Django
django==4.1.3
mysqlclient==2.0.3
djangorestframework==3.12.4
django-cors-headers==3.10.1
pytz==2021.3

# Storage
redis==4.0.2

# Celery
celery==5.2.2

# Server
uwsgi==2.0.20

# Datetime
arrow==1.2.3

# tencent cloud
tencentcloud-sdk-python==3.0.785

# OT
protobuf==3.19.5
opentelemetry-api==1.11.1
opentelemetry-sdk==1.11.1
opentelemetry-exporter-jaeger==1.11.1
opentelemetry-instrumentation==0.30b1
opentelemetry-instrumentation-django==0.30b1
opentelemetry-instrumentation-dbapi==0.30b1
opentelemetry-instrumentation-redis==0.30b1
opentelemetry-instrumentation-requests==0.30b1
opentelemetry-instrumentation-celery==0.30b1
opentelemetry-instrumentation-logging==0.30b1

# Log
python_json_logger==2.0.4

# RSA
pycryptodome==3.15.0

# Request
requests==2.27.1

# Profile
pyinstrument==4.4.0
